requirements.txt
mysql-connector-python
PySocks
